requirements_prod.txt
# Requirements actualizados para producción
Django==2.2.28
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
django-admin-generator==2.2.1
django-appconf==1.0.4
django-ckeditor==6.0.0
django-geoposition==0.3.0
django-image-cropping==1.5.0
django-import-export==2.5.0
django-js-asset==1.2.2
easy-thumbnails==2.7.1
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
Pillow>=8.0.0
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.5
xlrd==2.0.1
xlwt==1.3.0

# Servidor de producción
gunicorn==20.1.0
whitenoise==5.3.0
